Permissions added for Google Calander
Not sure where the pack-lock.json file came from... Appeared there magically when I was hitting random buttons. Might be a NPM thing, since I don't use yarn.

Still struggling to get the google calander API working. No success on my side so far. This commit don't have any of the mess I created to try to get something working. Intead of the mess, there's a comment block explaining the approach I took and why it wasn't working. on line 115 and onward in Setup.svelte.

Should google react google calander... Npm Google Calander...  After trying to do it with normal HTML/CSS/JS without svelte or whatever.

One thing also occured to me... The app's API key is exposed to the clients along with everything else in firebase.js. Maybe this is fine for now, to get the core of the project working. But if it takes off, bad idea to have it that way. Weirdos will do weird stuff with the api key!